consolidated piftools, final profiler bugfix
