+++ xrsdkit/models/__init__.py
@@ -11,7 +11,7 @@ from sklearn import preprocessing
 from .. import definitions as xrsdefs 
 from .regressor import Regressor
 from .classifier import Classifier
-from ..tools import primitives, profiler, citrination_tools
+from ..tools import primitives, profiler, piftools
 from ..system import System
 
 file_path = os.path.abspath(__file__)
@@ -180,7 +180,7 @@ def downsample_and_train(
         if True, the downsampling statistics and models will be
         saved in modeling_data/testing_data dir
     """
-    df = citrination_tools.get_data_from_Citrination(citrination_client,source_dataset_ids)
+    df = piftools.get_data_from_Citrination(citrination_client,source_dataset_ids)
     df_sample = downsample_by_group(df)
     train_from_dataframe(df_sample,train_hyperparameters,save_models,test)
